Changing user related code
