+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/controllers/UserController.java
@@ -8,7 +8,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
-@RequestMapping("users")
+@RequestMapping("/users")
 public class UserController {
     @Autowired
     private UserRepository userRepository;
