--- LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/models/dto/user.java
@@ -1,94 +1,0 @@
-package com.CraftyCoders.LaunchCash.models.dto;
-
-import com.CraftyCoders.LaunchCash.models.Profile;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotEmpty;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Entity
-public class user {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @NotEmpty
-    private String username;
-
-    @NotEmpty
-    private String hashedPassword;
-
-    @NotEmpty
-    private String name;
-
-    @NotEmpty
-    private String email;
-
-    private double balance;
-
-    private static final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-
-    @OneToOne(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Profile profile;
-
-    public user() {
-    }
-
-    public user(String username, String hashedPassword, String name, String email) {
-        this.username = username;
-        this.hashedPassword = hashedPassword;
-        this.name = name;
-        this.email = email;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getHashedPassword() {
-        return hashedPassword;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public double getBalance() {
-        return balance;
-    }
-
-    public void setBalance(double balance) {
-        this.balance = balance;
-    }
-
-    public void setHashedPassword(String hashedPassword) {
-        this.hashedPassword = encoder.encode(hashedPassword);
-    }
-    public boolean isMatchingPassword(String password) {
-        return encoder.matches(password, hashedPassword);
-    }
-
-}
