--- LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/repositories/userRepo.java
@@ -1,7 +1,0 @@
-package com.CraftyCoders.LaunchCash.repositories;
-
-import com.CraftyCoders.LaunchCash.models.dto.user;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface userRepo extends JpaRepository<user,Long> {
-}
