+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/services/UserService.java
@@ -1,0 +1,3 @@
+
+
+@Service
